% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Sub_boots_functions.R
\name{createBootsSibReg}
\alias{createBootsSibReg}
\title{Bootstrap and reformat raw data in the SibReg data format}
\usage{
createBootsSibReg(data.sibreg, boot.n = 100, plot.diagnostics = TRUE)
}
\arguments{
\item{boot.n}{some text}

\item{plot.diagnostics}{some text}

\item{dat.sibreg}{some text}
}
\description{
Bootstrap and reformat raw data in the SibReg data format
}
