% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_plotModelFit.R
\name{plotModelFit}
\alias{plotModelFit}
\title{Plot a model fit...}
\usage{
plotModelFit(
  fit.obj,
  options = list(plot.which = "all", age.which = "all", plot.add = FALSE),
  axis.label = "Abundance",
  fc.add = NULL,
  formula.show = TRUE,
  tracing = FALSE
)
}
\arguments{
\item{axis.label}{label for the value axis (in the app this is extracted from the data file, with user option to modify)}
}
\description{
Plot a model fit...
}
